Replace `Nl80211Cmd` with `Nl80211Command`
API breakage:
 * Replaced `Nl80211Cmd` by `Nl80211Command`.
 * Removed these `Nl80211Message` functions:
    * `new_interface_get()`
    * `new_station_get()`
    * `new_wiphy_get()`
    * `new_scan_get()`
